Moved tests classes into Tests folder
