+++ src/Atis.LinqToSql.UnitTest/Tests/DataManipulationQueryTests.cs
@@ -5,7 +5,7 @@ using System.Linq.Expressions;
 using System.Text;
 using System.Threading.Tasks;
 
-namespace Atis.LinqToSql.UnitTest
+namespace Atis.LinqToSql.UnitTest.Tests
 {
     [TestClass]
     public class DataManipulationQueryTests : TestBase
