+++ src/Atis.LinqToSql.UnitTest/Tests/TestBase.cs
@@ -9,7 +9,7 @@ using System.Linq.Expressions;
 using System.Text;
 using System.Threading.Tasks;
 
-namespace Atis.LinqToSql.UnitTest
+namespace Atis.LinqToSql.UnitTest.Tests
 {
     public abstract class TestBase
     {
